+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - stdlib: new function `lists.prepend` for prepending lists
 - fix: improved type switch error message
 - fix: improved stdlib not found error message
+- lsp: complete importable modules #35
+- lsp: field completions #35
 - chore: bump dependencies
 
 ## v0.0.17